refactor: Remove Field injection
closes #13
