+++ src/main/java/com/example/demo/DemoApplication.java
@@ -1,30 +1,11 @@
 package com.example.demo;
 
-import com.example.demo.post.PostRepository;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.servlet.ModelAndView;
 
-@Controller
 @SpringBootApplication
 public class DemoApplication {
-    //    TODO: Avoid field injection
-    @Autowired
-    private PostRepository postRepository;
-
     public static void main(String[] args) {
         SpringApplication.run(DemoApplication.class, args);
     }
-
-    @GetMapping("/")
-    ModelAndView home() {
-        ModelAndView modelAndView = new ModelAndView("index");
-        modelAndView.addObject("posts", postRepository.findAll());
-
-        return modelAndView;
-    }
-
 }
