+++ src/main/java/com/example/demo/post/PostService.java
@@ -1,0 +1,16 @@
+package com.example.demo.post;
+
+import org.springframework.stereotype.Service;
+
+@Service
+public class PostService {
+    private final PostRepository postRepository;
+
+    public PostService(PostRepository postRepository) {
+        this.postRepository = postRepository;
+    }
+
+    public Iterable<Post> findAll() {
+        return postRepository.findAll();
+    }
+}
